--- backend/src/main/java/com/woowacourse/zzimkkong/exception/AuthorizationException.java
@@ -1,16 +1,0 @@
-package com.woowacourse.zzimkkong.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class AuthorizationException extends RuntimeException {
-    private final HttpStatus httpStatus;
-
-    public AuthorizationException(final String message, final HttpStatus httpStatus) {
-        super(message);
-        this.httpStatus = httpStatus;
-    }
-
-    public HttpStatus getStatus() {
-        return httpStatus;
-    }
-}
